+++ platform-sdk/consensus-utility/src/main/java/module-info.java
@@ -5,6 +5,7 @@ import org.hiero.consensus.config.UtilityConfigurationExtension;
 module org.hiero.consensus.utility {
     exports org.hiero.consensus.monitoring;
     exports org.hiero.consensus.config;
+    exports org.hiero.consensus.constructable;
     exports org.hiero.consensus.crypto;
     exports org.hiero.consensus.event;
     exports org.hiero.consensus.event.validation;
@@ -22,10 +23,10 @@ module org.hiero.consensus.utility {
     requires transitive com.swirlds.config.api;
     requires transitive com.swirlds.metrics.api;
     requires transitive org.hiero.base.crypto;
+    requires transitive org.hiero.base.utility;
     requires transitive org.hiero.consensus.model;
     requires com.swirlds.logging;
     requires org.hiero.base.concurrent;
-    requires org.hiero.base.utility;
     requires org.hiero.consensus.concurrent;
     requires org.hiero.consensus.metrics;
     requires org.apache.logging.log4j;
