+++ platform-sdk/consensus-utility/src/main/java/org/hiero/consensus/constructable/ConstructableRegistration.java
@@ -1,12 +1,12 @@
 // SPDX-License-Identifier: Apache-2.0
-package com.swirlds.common.constructable;
+package org.hiero.consensus.constructable;
 
-import com.swirlds.common.merkle.utility.SerializableLong;
 import org.hiero.base.constructable.ClassConstructorPair;
 import org.hiero.base.constructable.ConstructableRegistry;
 import org.hiero.base.constructable.ConstructableRegistryException;
 import org.hiero.base.crypto.Hash;
 import org.hiero.base.crypto.SerializablePublicKey;
+import org.hiero.base.io.SerializableLong;
 import org.hiero.consensus.model.event.CesEvent;
 import org.hiero.consensus.model.node.NodeId;
