+++ platform-sdk/swirlds-common/src/main/java/module-info.java
@@ -3,7 +3,6 @@ module com.swirlds.common {
 
     /* Exported packages. This list should remain alphabetized. */
     exports com.swirlds.common.config;
-    exports com.swirlds.common.constructable;
     exports com.swirlds.common.context;
     exports com.swirlds.common.io.config;
     exports com.swirlds.common.io.exceptions;
@@ -14,15 +13,12 @@ module com.swirlds.common {
     exports com.swirlds.common.merkle.synchronization.task;
     exports com.swirlds.common.merkle.synchronization.utility;
     exports com.swirlds.common.merkle.synchronization.views;
-    exports com.swirlds.common.merkle.utility;
     exports com.swirlds.common.notification;
     exports com.swirlds.common.platform;
     exports com.swirlds.common.utility;
     exports com.swirlds.common.startup;
     exports com.swirlds.common.merkle.synchronization.stats;
 
-    opens com.swirlds.common.merkle.utility to
-            com.fasterxml.jackson.databind;
     opens com.swirlds.common.utility to
             com.fasterxml.jackson.databind;
     opens com.swirlds.common.io.utility to
@@ -38,7 +34,6 @@ module com.swirlds.common {
     requires transitive org.hiero.consensus.reconnect;
     requires transitive org.hiero.consensus.utility;
     requires com.swirlds.logging;
-    requires org.hiero.base.crypto;
     requires org.hiero.consensus.metrics;
     requires jdk.httpserver;
     requires jdk.management;
