+++ platform-sdk/swirlds-merkledb/build.gradle.kts
@@ -33,6 +33,7 @@ testModuleInfo {
     requires("org.apache.logging.log4j.core")
     requires("org.hiero.base.utility.test.fixtures")
     requires("org.hiero.consensus.model")
+    requires("org.hiero.consensus.utility")
     requires("org.junit.jupiter.api")
     requires("org.junit.jupiter.params")
     requires("org.mockito")
