+++ platform-sdk/swirlds-platform-core/src/main/java/com/swirlds/platform/state/iss/internal/DefaultIssHandler.java
@@ -2,7 +2,6 @@
 package com.swirlds.platform.state.iss.internal;
 
 import com.swirlds.common.context.PlatformContext;
-import com.swirlds.common.merkle.utility.SerializableLong;
 import com.swirlds.platform.components.common.output.FatalErrorConsumer;
 import com.swirlds.platform.scratchpad.Scratchpad;
 import com.swirlds.platform.state.iss.IssHandler;
@@ -10,6 +9,7 @@ import com.swirlds.platform.state.iss.IssScratchpad;
 import com.swirlds.platform.system.SystemExitCode;
 import edu.umd.cs.findbugs.annotations.NonNull;
 import java.util.Objects;
+import org.hiero.base.io.SerializableLong;
 import org.hiero.consensus.model.notification.IssNotification;
 import org.hiero.consensus.state.config.StateConfig;
