+++ platform-sdk/swirlds-state-impl/src/testFixtures/java/module-info.java
@@ -16,6 +16,7 @@ open module com.swirlds.state.impl.test.fixtures {
     requires com.swirlds.metrics.api;
     requires org.hiero.consensus.metrics;
     requires org.hiero.consensus.reconnect;
+    requires org.hiero.consensus.utility;
     requires org.junit.jupiter.api;
     requires static transitive com.github.spotbugs.annotations;
