+++ platform-sdk/swirlds-virtualmap/src/testFixtures/java/module-info.java
@@ -10,5 +10,6 @@ open module com.swirlds.virtualmap.test.fixtures {
     requires com.swirlds.virtualmap;
     requires org.hiero.base.utility;
     requires org.hiero.consensus.reconnect;
+    requires org.hiero.consensus.utility;
     requires org.junit.jupiter.api;
 }
